+++ src/fan_controller.rs
@@ -133,7 +133,8 @@ impl FanController {
             }
             SpeedCurve::Exponential => {
                 let exp = self.config.exp_pow;
-                ((temp - low_temp).pow(exp) as f32 / (high_temp - low_temp).pow(exp) as f32
+                (((temp - low_temp) as f32).powf(exp)
+                    / ((high_temp - low_temp) as f32).powf(exp)
                     * (self.max_speed - self.min_speed) as f32) as u32
                     + self.min_speed
             }
